:boom: Change command interface
