+++ deno.jsonc
@@ -15,6 +15,7 @@
   },
   "imports": {
     "@cliffy/ansi": "jsr:@cliffy/ansi@1.0.0-rc.7",
+    "@cliffy/command": "jsr:@cliffy/command@1.0.0-rc.7",
     "@cliffy/keypress": "jsr:@cliffy/keypress@1.0.0-rc.7",
     "@cliffy/table": "jsr:@cliffy/table@1.0.0-rc.7",
     "@eta-dev/eta": "jsr:@eta-dev/eta@3.5.0",
