+++ deno.lock
@@ -2,6 +2,8 @@
   "version": "5",
   "specifiers": {
     "jsr:@cliffy/ansi@1.0.0-rc.7": "1.0.0-rc.7",
+    "jsr:@cliffy/command@1.0.0-rc.7": "1.0.0-rc.7",
+    "jsr:@cliffy/flags@1.0.0-rc.7": "1.0.0-rc.7",
     "jsr:@cliffy/internal@1.0.0-rc.7": "1.0.0-rc.7",
     "jsr:@cliffy/keycode@1.0.0-rc.7": "1.0.0-rc.7",
     "jsr:@cliffy/keypress@1.0.0-rc.7": "1.0.0-rc.7",
@@ -12,6 +14,7 @@
     "jsr:@std/fmt@~1.0.2": "1.0.3",
     "jsr:@std/io@~0.224.9": "0.224.9",
     "jsr:@std/path@1.0.8": "1.0.8",
+    "jsr:@std/text@~1.0.7": "1.0.15",
     "jsr:@std/yaml@1.0.5": "1.0.5",
     "npm:@types/node@*": "22.13.8",
     "npm:xdg-basedir@5.1.0": "5.1.0"
@@ -24,6 +27,22 @@
         "jsr:@std/encoding",
         "jsr:@std/fmt@~1.0.2",
         "jsr:@std/io"
+      ]
+    },
+    "@cliffy/command@1.0.0-rc.7": {
+      "integrity": "1288808d7a3cd18b86c24c2f920e47a6d954b7e23cadc35c8cbd78f8be41f0cd",
+      "dependencies": [
+        "jsr:@cliffy/flags",
+        "jsr:@cliffy/internal",
+        "jsr:@cliffy/table",
+        "jsr:@std/fmt@~1.0.2",
+        "jsr:@std/text"
+      ]
+    },
+    "@cliffy/flags@1.0.0-rc.7": {
+      "integrity": "318d9be98f6a6417b108e03dec427dea96cdd41a15beb21d2554ae6da450a781",
+      "dependencies": [
+        "jsr:@std/text"
       ]
     },
     "@cliffy/internal@1.0.0-rc.7": {
@@ -60,6 +79,9 @@
     "@std/path@1.0.8": {
       "integrity": "548fa456bb6a04d3c1a1e7477986b6cffbce95102d0bb447c67c4ee70e0364be"
     },
+    "@std/text@1.0.15": {
+      "integrity": "91f5cc1e12779a3d95f1be34e763f9c28a75a078b7360e6fcaef0d8d9b1e3e7f"
+    },
     "@std/yaml@1.0.5": {
       "integrity": "71ba3d334305ee2149391931508b2c293a8490f94a337eef3a09cade1a2a2742"
     }
@@ -81,6 +103,7 @@
   "workspace": {
     "dependencies": [
       "jsr:@cliffy/ansi@1.0.0-rc.7",
+      "jsr:@cliffy/command@1.0.0-rc.7",
       "jsr:@cliffy/keypress@1.0.0-rc.7",
       "jsr:@cliffy/table@1.0.0-rc.7",
       "jsr:@eta-dev/eta@3.5.0",
