+++ src/widget.eta
@@ -3,7 +3,7 @@ _wk_widget() {
   zstyle -a ':wk:*' options opts || opts=()
 
   local res=''
-  res=$(<%= it.wk_path %> ${=opts} < $TTY 2>&1)
+  res=$(<%= it.wk_path %> run ${=opts} < $TTY 2>&1)
   local wk_exit=$?
 
   zle redisplay
